.ipynb_checkpoints/requirements-checkpoint.txt
streamlit
pandas
python-dotenv
requests
json
